requirements.txt
pandas==2.3.0
python-dotenv==1.2.1
openai==2.15.0
